fix: 배포오류 해결
