+++ src/app/_constants/feedback.ts
@@ -1,0 +1,17 @@
+export const goodReasons = [
+  "내게 꼭 맞는 이벤트를 추천해줘요",
+  "이벤트 신청/모집 과정이 간단해요",
+  "이벤트 정보를 자세히 확인할 수 있어요",
+  "알림으로 진행 과정을 확인할 수 있어 좋아요",
+  "대관한 영화관이 마음에 들어요",
+  "덕분에 잊지 못할 추억을 만들 수 있었어요",
+];
+
+export const badReasons = [
+  "원하는 이벤트를 찾기 불편해요",
+  "이벤트 신청/모집 과정이 어려워요",
+  "이벤트 상영까지 진행 과정이 너무 길어요",
+  "알림으로 진행 상황을 확인하는게 불편해요",
+  "대관 가능한 영화관이 너무 한정적이에요",
+  "이벤트가 취소되어 실망스러워요",
+];
