+++ src/app/_constants/index.ts
@@ -5,3 +5,4 @@ export { EVENT_CATEGORIES } from "./event-categories";
 export { PATH_IMAGES } from "./path-images";
 export { STATUS_MAP } from "./event-tab";
 export { EVENT_CREATE_TIME } from "./event-create";
+export { goodReasons, badReasons } from "./feedback";
